pyproject.toml
[project]
name = "myodoo-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
